Handle fs prefixes like "./" and "/"
This makes sure if either the tar file has these prefixes or the caller
uses these prefixes that we can find the file.

Note that it is technically possible for a tar to have an entry for both
"foo" and "./foo". In this case it is ambiguous what we should even do.
Since this is specifically for filesystems we'll take whichever entry
comes first in the tar.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>